Gotchas with Webpack Dev Server
webpack-dev-server does not create any files on your computer.
You can delete the dist folder on this repo, run webpack, and your app will show
up on the browser.

This is because the webpack-dev-server runs entirely in memory.  It calls webpack, and any files it emits is stored
as part of the webpack-dev-server instance.  This further emphasizes that dev-server is for DEVELOPMENT only.